+++ packages/braintree-utils/src/braintree-integration-service.ts
@@ -53,9 +53,10 @@ export default class BraintreeIntegrationService {
         this.braintreeScriptLoader.initialize(storeConfig);
     }
 
-    async getBraintreeConnect(cardId?: string) {
-        // TODO: should be removed after PayPal prepare stable Braintree SDK version with AXO implementation
-        window.localStorage.setItem('axoEnv', 'sandbox');
+    async getBraintreeConnect(cardId?: string, isTestModeEnabled?: boolean) {
+        if (isTestModeEnabled) {
+            window.localStorage.setItem('axoEnv', 'sandbox');
+        }
 
         if (!this.braintreeHostWindow.braintreeConnect) {
             const clientToken = this.getClientTokenOrThrow();