+++ packages/braintree-utils/src/index.ts
@@ -41,7 +41,6 @@ export {
     getPaypalCheckoutMock,
     getPayPalCheckoutCreatorMock,
     getBraintreeLocalPaymentMock,
-    getBraintree,
     getDeviceDataMock,
     getModuleCreatorMock,
     getGooglePayMock,